requirements.txt
openai>=1.0.0
python-dotenv
tenacity
pandas
matplotlib
seaborn
scikit-learn
jupyter
pytest
